Enhance handling iF and TRY.
Assume we have a test like this:

    *** Test Cases ***
    Example
        IF    True
            Fail    Not executed
        ELSE
            Log    Executed!
        END

The locator to match the `Log` keyword needed to be `1.2.1` where the
first `1` represents the IF/ELSE root object. This was changed so that
now the locator must be `2.1` i.e. the root object is totally omitted.

This new approach matches how IF and TRY look in data and in log
better than the earlier. The change is backwards incompatible, but
control structures haven't been properly supported prior to 4.0 so I
don't think this change alone warrants a new major release.